Deprecate R18v1, R66v1, R69v1; and promote the v2
