+++ packages/alfa-rules/src/deprecated.ts
@@ -1,0 +1,3 @@
+export { default as DR18 } from "./sia-dr18/rule";
+export { default as DR66 } from "./sia-dr66/rule";
+export { default as DR69 } from "./sia-dr69/rule";
