+++ packages/alfa-rules/src/experimental.ts
@@ -1,6 +1,3 @@
-export { default as ER18 } from "./sia-er18/rule";
-export { default as ER66 } from "./sia-er66/rule";
-export { default as ER69 } from "./sia-er69/rule";
 export { default as ER87 } from "./sia-er87/rule";
 
 export { default as R82 } from "./sia-r82/rule";
