+++ packages/alfa-rules/src/index.ts
@@ -7,8 +7,8 @@ export * from "./tags";
 import * as experimentalRules from "./experimental";
 export { experimentalRules };
 
-// import * as deprecatedRules from "./deprecated";
-// export { deprecatedRules };
+import * as deprecatedRules from "./deprecated";
+export { deprecatedRules };
 
 import * as rules from "./rules";
