+++ packages/alfa-rules/src/rules.ts
@@ -15,7 +15,7 @@ export { default as R14 } from "./sia-r14/rule";
 export { default as R15 } from "./sia-r15/rule";
 export { default as R16 } from "./sia-r16/rule";
 export { default as R17 } from "./sia-r17/rule";
-export { default as R18 } from "./sia-r18/rule";
+export { default as R18 } from "./sia-dr18/rule";
 export { default as R19 } from "./sia-r19/rule";
 export { default as R20 } from "./sia-r20/rule";
 export { default as R21 } from "./sia-r21/rule";
@@ -60,10 +60,10 @@ export { default as R62 } from "./sia-r62/rule";
 export { default as R63 } from "./sia-r63/rule";
 export { default as R64 } from "./sia-r64/rule";
 export { default as R65 } from "./sia-r65/rule";
-export { default as R66 } from "./sia-r66/rule";
+export { default as R66 } from "./sia-dr66/rule";
 export { default as R67 } from "./sia-r67/rule";
 export { default as R68 } from "./sia-r68/rule";
-export { default as R69 } from "./sia-r69/rule";
+export { default as R69 } from "./sia-dr69/rule";
 export { default as R70 } from "./sia-r70/rule";
 export { default as R71 } from "./sia-r71/rule";
 export { default as R72 } from "./sia-r72/rule";
