+++ packages/alfa-rules/src/sia-r18/rule.ts
@@ -12,7 +12,7 @@ import * as aria from "@siteimprove/alfa-aria";
 
 import { expectation } from "../common/act/expectation";
 
-import { Scope } from "../tags";
+import { Scope, Version } from "../tags";
 
 const { hasRole, isIncludedInTheAccessibilityTree } = DOM;
 const { hasDisplaySize, hasInputType } = Element;
@@ -21,7 +21,7 @@ const { test, property } = Predicate;
 export default Rule.Atomic.of<Page, Attribute>({
   uri: "https://alfa.siteimprove.com/rules/sia-r18",
   requirements: [Technique.of("ARIA5")],
-  tags: [Scope.Component],
+  tags: [Scope.Component, Version.of(2)],
   evaluate({ device, document }) {
     const global = Set.from(Role.of("roletype").supportedAttributes);
 
@@ -38,6 +38,9 @@ export default Rule.Atomic.of<Page, Attribute>({
       },
 
       expectations(target) {
+        // Since the attribute was found on a element, it has a owner.
+        const owner = target.owner.getUnsafe();
+
         return {
           1: expectation(
             global.has(target.name as aria.Attribute.Name) ||
@@ -45,12 +48,21 @@ export default Rule.Atomic.of<Page, Attribute>({
                 hasRole(device, (role) =>
                   role.isAttributeSupported(target.name as aria.Attribute.Name)
                 ),
-                // Since the attribute was found on a element, it has a owner.
-                target.owner.getUnsafe()
+                owner
               ) ||
               ariaHtmlAllowed(target),
             () => Outcomes.IsAllowed,
             () => Outcomes.IsNotAllowed
+          ),
+          2: expectation(
+            test(
+              hasRole(device, (role) =>
+                role.isAttributeProhibited(target.name as aria.Attribute.Name)
+              ),
+              owner
+            ),
+            () => Outcomes.IsProhibited,
+            () => Outcomes.IsNotProhibited
           ),
         };
       },
@@ -105,6 +117,18 @@ export namespace Outcomes {
   export const IsNotAllowed = Err.of(
     Diagnostic.of(
       `The attribute is not allowed for the element on which it is specified`
+    )
+  );
+
+  export const IsProhibited = Err.of(
+    Diagnostic.of(
+      `The attribute is prohibited for the element on which it is specified`
+    )
+  );
+
+  export const IsNotProhibited = Ok.of(
+    Diagnostic.of(
+      `The attribute is not prohibited for the element on which it is specified`
     )
   );
 }
