+++ packages/alfa-rules/src/sia-r66/rule.ts
@@ -1,19 +1,25 @@
 import { Rule } from "@siteimprove/alfa-act";
-import { Text } from "@siteimprove/alfa-dom";
+import { Element, Text } from "@siteimprove/alfa-dom";
 import { Criterion } from "@siteimprove/alfa-wcag";
 import { Page } from "@siteimprove/alfa-web";
 
+import { Group } from "../common/act/group";
 import { Question } from "../common/act/question";
 import { nonDisabledTexts } from "../common/applicability/non-disabled-texts";
 
-import { hasSufficientContrast } from "../common/expectation/contrast";
+import { hasSufficientContrastExperimental } from "../common/expectation/contrast";
 
-import { Scope } from "../tags";
+import { Scope, Version } from "../tags";
 
-export default Rule.Atomic.of<Page, Text, Question.Metadata>({
+export default Rule.Atomic.of<
+  Page,
+  Text,
+  Question.Metadata,
+  Text | Group<Element>
+>({
   uri: "https://alfa.siteimprove.com/rules/sia-r66",
   requirements: [Criterion.of("1.4.6")],
-  tags: [Scope.Component],
+  tags: [Scope.Component, Version.of(2)],
   evaluate({ device, document }) {
     return {
       applicability() {
@@ -21,7 +27,7 @@ export default Rule.Atomic.of<Page, Text, Question.Metadata>({
       },
 
       expectations(target) {
-        return hasSufficientContrast(target, device, 4.5, 7);
+        return hasSufficientContrastExperimental(target, device, 4.5, 7);
       },
     };
   },
