+++ packages/alfa-rules/test/sia-r18/rule.spec.tsx
@@ -18,6 +18,7 @@ test(`evaluate() passes input with type password field and aria-required state`,
   t.deepEqual(await evaluate(R18, { document }), [
     passed(R18, target.attribute("aria-required").getUnsafe(), {
       1: Outcomes.IsAllowed,
+      2: Outcomes.IsNotProhibited,
     }),
   ]);
 });
@@ -30,6 +31,7 @@ test(`evaluate() passes a button with aria-pressed state`, async (t) => {
   t.deepEqual(await evaluate(R18, { document }), [
     passed(R18, target.attribute("aria-pressed").getUnsafe(), {
       1: Outcomes.IsAllowed,
+      2: Outcomes.IsNotProhibited,
     }),
   ]);
 });
@@ -46,6 +48,7 @@ test(`evaluate() passes a div element with button role, and an aria-pressed stat
   t.deepEqual(await evaluate(R18, { document }), [
     passed(R18, target.attribute("aria-pressed").getUnsafe(), {
       1: Outcomes.IsAllowed,
+      2: Outcomes.IsNotProhibited,
     }),
   ]);
 });
@@ -58,6 +61,7 @@ test(`evaluate() passes a div element with aria busy state`, async (t) => {
   t.deepEqual(await evaluate(R18, { document }), [
     passed(R18, target.attribute("aria-busy").getUnsafe(), {
       1: Outcomes.IsAllowed,
+      2: Outcomes.IsNotProhibited,
     }),
   ]);
 });
@@ -74,6 +78,7 @@ test(`evaluate() passes a div element with button role, and an aria-label attrib
   t.deepEqual(await evaluate(R18, { document }), [
     passed(R18, target.attribute("aria-label").getUnsafe(), {
       1: Outcomes.IsAllowed,
+      2: Outcomes.IsNotProhibited,
     }),
   ]);
 });
@@ -90,6 +95,7 @@ test(`evaluate() passes a div element with checkbox role, and an aria-checked st
   t.deepEqual(await evaluate(R18, { document }), [
     passed(R18, target.attribute("aria-checked").getUnsafe(), {
       1: Outcomes.IsAllowed,
+      2: Outcomes.IsNotProhibited,
     }),
   ]);
 });
@@ -106,9 +112,11 @@ test(`evaluate() passes a div element with checkbox role, and an aria-controls s
   t.deepEqual(await evaluate(R18, { document }), [
     passed(R18, target.attribute("aria-controls").getUnsafe(), {
       1: Outcomes.IsAllowed,
+      2: Outcomes.IsNotProhibited,
     }),
     passed(R18, target.attribute("aria-expanded").getUnsafe(), {
       1: Outcomes.IsAllowed,
+      2: Outcomes.IsNotProhibited,
     }),
   ]);
 });
@@ -126,9 +134,11 @@ test(`evaluate() passes a div element with checkbox role, and both
   t.deepEqual(await evaluate(R18, { document }), [
     passed(R18, target.attribute("aria-controls").getUnsafe(), {
       1: Outcomes.IsAllowed,
+      2: Outcomes.IsNotProhibited,
     }),
     passed(R18, target.attribute("aria-expanded").getUnsafe(), {
       1: Outcomes.IsAllowed,
+      2: Outcomes.IsNotProhibited,
     }),
   ]);
 });
@@ -146,9 +156,11 @@ test(`evaluate() passes a div element with checkbox role, with both
   t.deepEqual(await evaluate(R18, { document }), [
     passed(R18, target.attribute("aria-expanded").getUnsafe(), {
       1: Outcomes.IsAllowed,
+      2: Outcomes.IsNotProhibited,
     }),
     passed(R18, target.attribute("aria-controls").getUnsafe(), {
       1: Outcomes.IsAllowed,
+      2: Outcomes.IsNotProhibited,
     }),
   ]);
 });
@@ -165,6 +177,7 @@ test(`evaluate() passes a button element with none role and aria-pressed`, async
   t.deepEqual(await evaluate(R18, { document }), [
     passed(R18, target.attribute("aria-pressed").getUnsafe(), {
       1: Outcomes.IsAllowed,
+      2: Outcomes.IsNotProhibited,
     }),
   ]);
 });
@@ -177,6 +190,20 @@ test(`evaluate() fails a button with aria-sort state, and no property`, async (t
   t.deepEqual(await evaluate(R18, { document }), [
     failed(R18, target.attribute("aria-sort").getUnsafe(), {
       1: Outcomes.IsNotAllowed,
+      2: Outcomes.IsNotProhibited,
+    }),
+  ]);
+});
+
+test(`evaluate() fails a prohibited attribute`, async (t) => {
+  const target = <div aria-label="foo">bar</div>;
+
+  const document = h.document([target]);
+
+  t.deepEqual(await evaluate(R18, { document }), [
+    failed(R18, target.attribute("aria-label").getUnsafe(), {
+      1: Outcomes.IsAllowed,
+      2: Outcomes.IsProhibited,
     }),
   ]);
 });
