+++ packages/alfa-rules/tsconfig.json
@@ -51,9 +51,9 @@
     "src/experimental.ts",
     "src/index.ts",
     "src/rules.ts",
-    "src/sia-er18/rule.ts",
-    "src/sia-er66/rule.ts",
-    "src/sia-er69/rule.ts",
+    "src/sia-dr18/rule.ts",
+    "src/sia-dr66/rule.ts",
+    "src/sia-dr69/rule.ts",
     "src/sia-er87/rule.ts",
     "src/sia-r1/rule.ts",
     "src/sia-r10/rule.ts",
@@ -160,9 +160,9 @@
     "test/common/outcome.ts",
     "test/common/dom/get-colors.spec.tsx",
     "test/common/predicate/is-at-the-start.spec.tsx",
-    "test/sia-er18/rule.spec.tsx",
-    "test/sia-er66/rule.spec.tsx",
-    "test/sia-er69/rule.spec.tsx",
+    "test/sia-dr18/rule.spec.tsx",
+    "test/sia-dr66/rule.spec.tsx",
+    "test/sia-dr69/rule.spec.tsx",
     "test/sia-er87/rule.spec.tsx",
     "test/sia-r1/rule.spec.tsx",
     "test/sia-r2/rule.spec.tsx",
